feat(dependencies): add router
vue-routerを追加した。
addページを作成
